% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.dos.numeros.R
\name{suma.dos.numeros}
\alias{suma.dos.numeros}
\title{Suma dos números}
\usage{
suma.dos.numeros(x, y)
}
\arguments{
\item{x}{Un número. El primer número a sumar.}

\item{y}{Un número. El segundo número a sumar.}
}
\value{
La suma de los dos números.
}
\description{
Esta función toma dos números y devuelve su suma.
}
